refactor: move kiosk control panel to ui
